requirements.txt
Django~=2.2
psycopg2-binary==2.8.1
djangorestframework~=3.9.2
redis~=3.2.1
celery~=4.1.1
pytz>=2017.3
django-cors-headers
djangorestframework-jwt
djangorestframework-camel-case~=1.0.3
phonenumbers
Pillow
django-phonenumber-field
django-debug-toolbar~=1.9.1
uwsgi
django_extensions
pyparsing
pydot
requests~=2.20
beautifulsoup4~=4.7.1
drf-yasg~=1.16.0